SecureNotes/requirements.txt
Flask
firebase-admin
pycryptodome
